requirements.txt
# Core dependencies
kafka-python==2.0.2
pydantic>=2.8.0
pyyaml==6.0.1
jsonschema==4.20.0

# LLM Provider integrations
openai==1.3.0
anthropic==0.7.0
google-cloud-aiplatform==1.38.0

# Utilities
asyncio-mqtt==0.16.1
aiofiles==23.2.1
structlog==23.2.0
prometheus-client==0.19.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
